requirements.txt
gunicorn==19.5.0
Django==1.9.6
django_compressor==2.0
django-static-precompiler==1.4.0
django-haystack==2.5.dev1
django-http-proxy==0.4.3
dj-database-url==0.4.1
lxml==3.6.0
mysqlclient==1.3.7
pysolr==3.5.0
whitenoise==3.1
